clarify the file_contexts.subs_dist configuration file usage from Guido Trentalancia
Add a comment at the top of the configuration file file_contexts.subs_dist
to clarify that it performs aliasing and not substitutions in the
strict sense of the word.

A name change might be considered too, if it proves to lead to further
confusion.

There might be pieces of documentation that could benefit from similar
considerations.

Also note that a specific manual page is missing.