Display a list of useful policy modules from the .fc files
It happens that refpolicy users want to reduce the number of modules that
are used on their systems to a minimal set. Doing so by hand is prone to
errors, because one might forget that a piece of software is installed
on the target system, in which case files and processes will be
mislabeled.

In order to help detecting such issues, I wrote a script a few years
ago which converted the regular expressions in .fc files to glob()
patterns (aka. the patterns used by several shells to match paths). As
this issue came again recently on IRC (#selinux channel), I cleaned up
the script and suggest adding it to the main repository.

Signed-off-by: Nicolas Iooss <nicolas.iooss@m4x.org>