Remove systemd-analyze domain and executable type
The systemd-analyze domain has no unique permissions, and general usage seems to be to simply run in the caller domain.  Running in the caller domain makes sense because the access required is generally the access to read the information that systemd-analyze reads.  If the calling domain should be able to read the information it makes sense to grant it directly.

Signed-off-by: Daniel Burgener <dburgener@linux.microsoft.com>