portage: permit to build on tmpfs
My portage PORTAGE_TMPDIR directory is on a tmpfs, but emerging openssl fail with:
Can not execute files in /var/tmp/portage
This is due to missing rules permiting portage to execute files in
tmpfs.
After this error, some relabel/manage rules was still missing, so this
commit add them also.

Signed-off-by: Corentin LABBE <clabbe.montjoie@gmail.com>