add permissions for the systemd.offline-updates symlink
Adds support the "magic symlink" at /system-update or
/etc/system-update. The package manager is supposed to create this
link, a system generator checks for it, then the package manager needs
to read it and then delete it to clean up after itself. The initrc_t
permissions are because systemd provides a unit to forcibly cleanup in
case the update program doesn't for some reason.

Signed-off-by: Michael Snook <sgtsnookums@gmail.com>